move orphaned components/consts into their only consumers
