--- static/app/views/alerts/rules/metric/details/relatedIssuesNotAvailable.tsx
@@ -1,2 +1,0 @@
-export const RELATED_ISSUES_BOOLEAN_QUERY_ERROR =
-  'Error parsing search query: Boolean statements containing "OR" or "AND" are not supported in this search';
