+++ static/app/views/dashboards/widgetBuilder/components/widgetOnDemandQueryWarning.tsx
@@ -1,4 +1,10 @@
-import {OnDemandWarningIcon} from 'sentry/components/alerts/onDemandMetricAlert';
+import type React from 'react';
+import styled from '@emotion/styled';
+
+import {Tooltip} from '@sentry/scraps/tooltip';
+
+import {IconWarning} from 'sentry/icons';
+import {SvgIcon} from 'sentry/icons/svgIcon';
 import {tct} from 'sentry/locale';
 import {isOnDemandQueryString} from 'sentry/utils/onDemandMetrics';
 import {hasOnDemandMetricWidgetFeature} from 'sentry/utils/onDemandMetrics/features';
@@ -49,3 +55,26 @@ export function WidgetOnDemandQueryWarning(props: {
     />
   );
 }
+
+export function OnDemandWarningIcon({
+  msg,
+  isHoverable,
+  variant = 'muted',
+}: {
+  msg: React.ReactNode;
+  isHoverable?: boolean;
+  variant?: 'primary' | 'warning' | 'danger' | 'muted';
+}) {
+  return (
+    <Tooltip skipWrapper title={msg} isHoverable={isHoverable}>
+      <HoverableIconWarning variant={variant} />
+    </Tooltip>
+  );
+}
+
+const HoverableIconWarning = styled(IconWarning)`
+  min-width: ${() => SvgIcon.ICON_SIZES.sm};
+  &:hover {
+    cursor: pointer;
+  }
+`;
